Initial setup
The idea is to make a driver for the MMMPI's high speed 16550-style serial
ports. I'm starting from the s16550 driver built by Randy Wilson, used by
the Fast232 device sold by Rick Uland in the '90s. The s16550 driver was,
kinda-sorta, based on the SACIA driver built by Bruce Isted for the 6551
ACIA chip used in the Tandy RS-232 and Modem Paks (and almost every other
CoCo serial device). How's that for an incestuous family tree?

The 16550 is a very different beast from the 6551 -- the 16550 is actually
an upgraded 8250 (used a lot on PCs of the day), so despite the similar
names they're nothing alike.

The assembly source is actually a really bad disassembly that assembles, but
only works on a Hitachi 6309 CPU, in native mode, and only with the Fast232
device. So there's that. We'll see what we can do.

I have made a defs file, "my8250.d", that includes definitions for 8250 and
derived chips. Should be useful for figuring out what's what.