Info drop from L. Curtis Boyle
Curtis sent me a copy of the disassembly he's been working on of the "short"
(feature-disabled) version of the s16550 driver. I'll be using it to help
decipher what the hell's going on in the big one. :)

He also sent me a document he got from Rick describing some of the
specifics of the old driver.