Start documenting code
OK, so the first $1D reserved bytes in the static data area are for the
descriptor. I also figured out that the reservation at the end of the
static data area are the _transmit_ buffer, and what each byte actually
means.

I've found a lot of the register use in GetStat/SetStat, which is important
because that's where the 6809/6309 compatibility problems will be found.

Why? Because any interrupt (which system calls are) pushes all the CPU
registers onto the S stack, and pulls them back off when the interrupt
returns. This is all well and good, but on a 6309 in native mode (which
NitrOS9/6309 is in), the registers are pushed and pulled differently for
interrupts. The 6809 interrupt pushes, in order: PC, U, Y, X, DP, B, A, CC.
However, in native mode a 6309 pushes: PC, U, Y, X, DP, F, E, B, A, CC.

So what? you may ask. Well, if in a driver's GetStat function you want to
return values to the program that called I$GetStt (and you do), you actually
have to put those returned values into a different place depending on what
kind of CPU you find yourself running on. If you used the 6809 offset for Y
on a 6309, you'll actually be returning the "Y" value in the X register.
Even worse, if you try returning something in X, you'll be setting the
caller's direct page to a new value, probably causing all sorts of
"interesting" behavior.

So, drivers that implement GetStat mostly need to be CPU-specific, and they
need to use the R$[reg] symbols to refer to the caller's registers.

I've also made good progress on locating some of the "magic" numbers used.